backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Bcrypt==1.0.1